A reported change needed for TBase
TBase diverges from PG 10. Reportedly the offset of
t_infomask in struct HeapTupleHeaderData is different.
(Perhaps it is simply built with different compiler
options for structure packing?)

I hope the compiler is not giving up after reporting one
error. There are other members of HeapTupleHeaderData that
follow t_infomask, so it is possible their offsets will
also change. It would be irritating for the compiler to
only report another mismatch after a preceding one is
fixed.

In any case, this commit demonstrates how to accommodate
such a change. A constant deemed to have been stable over
many PG versions might simply be defined in the Java code
and use a CONFIRM... macro in ModelConstants.c to detect the
unexpected case where it changes.

If a constant turns out to be less constant than that, then
the approach has to change. That means deleting the simple
Java definition of the constant (here found in
TupleTableSlotImpl.java) and the matching CONFIRMOFFSET
in ModelConstants.c, and then adding four things, three
of them in ModelConstants.java: 1) an assigned IDX_...
index for a new constant slot; 2) a public static final
declaration for the constant; 3) a statement within the
static initializer that assigns the value with a
checked(...) call.

The fourth thing to add is a line in ModelConstants.c
that stores the needed value into the assigned slot.

Because TupleTableSlotImpl.java used to define that
offset locally, and now doesn't, it could be necessary
to add a static import from ModelConstants.java. But
in the case of TupleTableSlotImpl, there is already
an import static ModelConstants.* so no change is
needed there.